fix: getCategoryName invocation
